+++ packages/components/psammead-brand/src/index.jsx
@@ -1,9 +1,12 @@
 import React, { Fragment } from 'react';
 import styled from 'styled-components';
-import { string, number, node, shape } from 'prop-types';
+import { string, number, node, shape, bool } from 'prop-types';
 import { C_POSTBOX, C_WHITE } from '@bbc/psammead-styles/colours';
 import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
-import { GEL_GROUP_3_SCREEN_WIDTH_MIN } from '@bbc/gel-foundations/breakpoints';
+import {
+  GEL_GROUP_3_SCREEN_WIDTH_MIN,
+  GEL_GROUP_5_SCREEN_WIDTH_MIN,
+} from '@bbc/gel-foundations/breakpoints';
 import {
   GEL_SPACING_HLF,
   GEL_SPACING,
@@ -19,6 +22,13 @@ const PADDING_AROUND_SVG_BELOW_600PX = 32;
 const conditionallyRenderHeight = (svgHeight, padding) =>
   svgHeight ? `height: ${(svgHeight + padding) / 16}rem` : '';
 
+const TRANSPARENT_BORDER = `0.0625rem solid transparent`;
+
+const SvgWrapper = styled.div`
+  max-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN};
+  margin: 0 auto;
+`;
+
 const Banner = styled.div`
   background-color: ${C_POSTBOX};
   ${({ svgHeight }) =>
@@ -31,6 +41,8 @@ const Banner = styled.div`
       conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_ABOVE_600PX)};
     padding: 0 ${GEL_SPACING_DBL};
   }
+  border-top: ${({ borderTop }) => borderTop && TRANSPARENT_BORDER};
+  border-bottom: ${({ borderBottom }) => borderBottom && TRANSPARENT_BORDER};
 `;
 
 const brandWidth = (minWidth, maxWidth) => `
@@ -71,7 +83,32 @@ const BrandSvg = styled.svg`
   /* stylelint-enable */
 `;
 
-const StyledBrand = ({ brandName, svgHeight, svg, maxWidth, minWidth }) => (
+const LocalisedBrandName = ({ product, serviceLocalisedName }) =>
+  serviceLocalisedName ? (
+    <Fragment>
+      <span lang="en-GB">{product}</span>, {serviceLocalisedName}
+    </Fragment>
+  ) : (
+    product
+  );
+
+LocalisedBrandName.propTypes = {
+  product: string.isRequired,
+  serviceLocalisedName: string,
+};
+
+LocalisedBrandName.defaultProps = {
+  serviceLocalisedName: null,
+};
+
+const StyledBrand = ({
+  product,
+  serviceLocalisedName,
+  svgHeight,
+  svg,
+  maxWidth,
+  minWidth,
+}) => (
   <Fragment>
     {svg && (
       <Fragment>
@@ -87,14 +124,20 @@ const StyledBrand = ({ brandName, svgHeight, svg, maxWidth, minWidth }) => (
         >
           {svg.group}
         </BrandSvg>
-        <VisuallyHiddenText>{brandName}</VisuallyHiddenText>
+        <VisuallyHiddenText>
+          <LocalisedBrandName
+            product={product}
+            serviceLocalisedName={serviceLocalisedName}
+          />
+        </VisuallyHiddenText>
       </Fragment>
     )}
   </Fragment>
 );
 
 const brandProps = {
-  brandName: string.isRequired,
+  product: string.isRequired,
+  serviceLocalisedName: string,
   maxWidth: number.isRequired,
   minWidth: number.isRequired,
   svgHeight: number.isRequired,
@@ -110,37 +153,47 @@ const brandProps = {
 
 StyledBrand.propTypes = brandProps;
 
-const Brand = ({ brandName, svgHeight, minWidth, maxWidth, svg, url }) => (
-  <Banner svgHeight={svgHeight}>
-    {url ? (
-      <StyledLink href={url} maxWidth={maxWidth} minWidth={minWidth}>
-        <StyledBrand
-          brandName={brandName}
-          svg={svg}
-          maxWidth={maxWidth}
-          minWidth={minWidth}
-          svgHeight={svgHeight}
-        />
-      </StyledLink>
-    ) : (
-      <StyledBrand
-        brandName={brandName}
-        svg={svg}
-        svgHeight={svgHeight}
-        minWidth={minWidth}
-        maxWidth={maxWidth}
-      />
-    )}
-  </Banner>
-);
+StyledBrand.defaultProps = {
+  serviceLocalisedName: null,
+};
+
+const Brand = props => {
+  const { svgHeight, maxWidth, minWidth, url, borderTop, borderBottom } = props;
+
+  return (
+    <Banner
+      svgHeight={svgHeight}
+      borderTop={borderTop}
+      borderBottom={borderBottom}
+    >
+      {url ? (
+        <SvgWrapper>
+          <StyledLink href={url} maxWidth={maxWidth} minWidth={minWidth}>
+            <StyledBrand {...props} />
+          </StyledLink>
+        </SvgWrapper>
+      ) : (
+        <SvgWrapper>
+          <StyledBrand {...props} />
+        </SvgWrapper>
+      )}
+    </Banner>
+  );
+};
 
 Brand.defaultProps = {
   url: null,
+  serviceLocalisedName: null,
+  borderTop: false,
+  borderBottom: false,
 };
 
 Brand.propTypes = {
   ...brandProps,
   url: string,
+  serviceLocalisedName: string,
+  borderTop: bool,
+  borderBottom: bool,
 };
 
 export default Brand;