+++ packages/components/psammead-brand/src/index.stories.jsx
@@ -1,5 +1,11 @@
 import React from 'react';
-import { select, number, withKnobs } from '@storybook/addon-knobs';
+import {
+  select,
+  number,
+  text,
+  withKnobs,
+  boolean,
+} from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import * as svgs from '@bbc/psammead-assets/svgs';
 import { dirDecorator } from '@bbc/psammead-storybook-helpers';
@@ -13,14 +19,27 @@ const inputs = () => {
     .map(key => key.charAt(0).toUpperCase() + key.slice(1));
 
   const svgChoice = select('Service SVG', options, 'News').toLowerCase();
+  const productInput = text('Product', 'BBC News');
+  const serviceLocalisedNameInput = text('Localised service name', 'Yoruba');
   const svgRatio = svgs[svgChoice].ratio;
   const svgMaxHeight = 24;
   const svgMinHeight = 16;
   const minWidthInput = number('minimum svg width', svgRatio * svgMinHeight);
   const maxWidthInput = number('maximum svg width', svgRatio * svgMaxHeight);
   const svgHeightInput = number('desired height svg', svgMaxHeight);
+  const borderBottom = boolean('Border Bottom', false);
+  const borderTop = boolean('Border Top', false);
 
-  return { svgChoice, svgHeightInput, minWidthInput, maxWidthInput };
+  return {
+    productInput,
+    serviceLocalisedNameInput,
+    svgChoice,
+    svgHeightInput,
+    minWidthInput,
+    maxWidthInput,
+    borderTop,
+    borderBottom,
+  };
 };
 
 storiesOf('Components|Brand', module)
@@ -30,19 +49,26 @@ storiesOf('Components|Brand', module)
     'without brand link',
     () => {
       const {
+        productInput,
+        serviceLocalisedNameInput,
         svgHeightInput,
         minWidthInput,
         maxWidthInput,
         svgChoice,
+        borderBottom,
+        borderTop,
       } = inputs();
 
       return (
         <Brand
-          brandName="Default Brand Name"
+          product={productInput}
+          serviceLocalisedName={serviceLocalisedNameInput}
           svgHeight={svgHeightInput}
           minWidth={minWidthInput}
           maxWidth={maxWidthInput}
           svg={svgs[svgChoice]}
+          borderBottom={borderBottom}
+          borderTop={borderTop}
         />
       );
     },
@@ -52,20 +78,27 @@ storiesOf('Components|Brand', module)
     'with brand link',
     () => {
       const {
+        productInput,
+        serviceLocalisedNameInput,
         svgHeightInput,
         minWidthInput,
         maxWidthInput,
         svgChoice,
+        borderBottom,
+        borderTop,
       } = inputs();
 
       return (
         <Brand
-          brandName="Default Brand Name"
+          product={productInput}
+          serviceLocalisedName={serviceLocalisedNameInput}
           svgHeight={svgHeightInput}
           minWidth={minWidthInput}
           maxWidth={maxWidthInput}
           svg={svgs[svgChoice]}
-          url="https://bbc.com/news"
+          url="https://www.bbc.com/news"
+          borderBottom={borderBottom}
+          borderTop={borderTop}
         />
       );
     },