+++ packages/components/psammead-sitewide-links/src/List/index.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import styled from 'styled-components';
 import { string, arrayOf, shape } from 'prop-types';
-import nanoid from 'nanoid';
 import { C_SHADOW } from '@bbc/psammead-styles/colours';
 import { GEL_SPACING, GEL_SPACING_DBL } from '@bbc/gel-foundations/spacings';
 import {
+  GEL_GROUP_0_SCREEN_WIDTH_MAX,
+  GEL_GROUP_1_SCREEN_WIDTH_MIN,
   GEL_GROUP_2_SCREEN_WIDTH_MAX,
   GEL_GROUP_3_SCREEN_WIDTH_MIN,
   GEL_GROUP_3_SCREEN_WIDTH_MAX,
@@ -15,6 +16,11 @@ import {
 
 import Link from '../Link';
 
+// Gets the number of grid rows, taking into account the
+// first-child in the grid being separate, on its own row.
+const getRowCount = (children, columns) =>
+  Math.ceil((React.Children.count(children) - 1) / columns) + 1;
+
 const StyledList = styled.ul`
   border-bottom: 1px solid ${C_SHADOW};
   display: grid;
@@ -22,25 +28,40 @@ const StyledList = styled.ul`
   list-style-type: none;
   margin: 0;
   padding: 0 0 ${GEL_SPACING};
-  @media (max-width: ${GEL_GROUP_2_SCREEN_WIDTH_MAX}) {
+  @media (max-width: ${GEL_GROUP_0_SCREEN_WIDTH_MAX}) {
+    grid-auto-flow: row;
+  }
+  @media (min-width: ${GEL_GROUP_1_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_2_SCREEN_WIDTH_MAX}) {
     grid-column-gap: ${GEL_SPACING};
     grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(4, auto);
+    grid-template-rows: repeat(
+      ${({ children }) => getRowCount(children, 2)},
+      auto
+    );
   }
   @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_3_SCREEN_WIDTH_MAX}) {
     grid-column-gap: ${GEL_SPACING_DBL};
     grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: repeat(3, auto);
+    grid-template-rows: repeat(
+      ${({ children }) => getRowCount(children, 3)},
+      auto
+    );
   }
   @media (min-width: ${GEL_GROUP_4_SCREEN_WIDTH_MIN}) and (max-width: ${GEL_GROUP_4_SCREEN_WIDTH_MAX}) {
     grid-column-gap: ${GEL_SPACING_DBL};
     grid-template-columns: repeat(4, 1fr);
-    grid-template-rows: repeat(3, auto);
+    grid-template-rows: repeat(
+      ${({ children }) => getRowCount(children, 4)},
+      auto
+    );
   }
   @media (min-width: ${GEL_GROUP_5_SCREEN_WIDTH_MIN}) {
     grid-column-gap: ${GEL_SPACING_DBL};
     grid-template-columns: repeat(5, 1fr);
-    grid-template-rows: repeat(3, auto);
+    grid-template-rows: repeat(
+      ${({ children }) => getRowCount(children, 5)},
+      auto
+    );
   }
   > li:first-child {
     border-bottom: 1px solid ${C_SHADOW};
@@ -63,7 +84,7 @@ const StyledListItem = styled.li`
 const List = ({ links }) => (
   <StyledList role="list">
     {links.map(link => (
-      <StyledListItem key={nanoid()} role="listitem">
+      <StyledListItem key={link.text} role="listitem">
         <Link text={link.text} href={link.href} />
       </StyledListItem>
     ))}