+++ packages/components/psammead-story-promo-list/src/index.test.jsx
@@ -14,7 +14,7 @@ const Info = (
   </Fragment>
 );
 
-describe('StoryPromo', () => {
+describe('StoryPromo list', () => {
   shouldMatchSnapshot(
     'should render correctly',
     <StoryPromoUl>