+++ packages/components/psammead-story-promo/src/index.test.jsx
@@ -2,16 +2,20 @@ import React, { Fragment } from 'react';
 import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import { latin } from '@bbc/gel-foundations/scripts';
 import MediaIndicator from '@bbc/psammead-media-indicator';
+import { render } from '@testing-library/react';
 import StoryPromo, { Headline, Summary, Link } from './index';
 
 const Image = <img src="https://foobar.com/image.png" alt="Alt text" />;
 
-const Info = (
+// eslint-disable-next-line react/prop-types
+const Info = ({ topStory }) => (
   <Fragment>
-    <Headline script={latin}>
+    <Headline script={latin} topStory={topStory}>
       <Link href="https://www.bbc.co.uk/news">The headline of the promo</Link>
     </Headline>
-    <Summary script={latin}>The summary of the promo</Summary>
+    <Summary script={latin} topStory={topStory}>
+      The summary of the promo
+    </Summary>
     <time>12 March 2019</time>
   </Fragment>
 );
@@ -27,13 +31,66 @@ const mediaInfo = (
 describe('StoryPromo', () => {
   shouldMatchSnapshot(
     'should render correctly',
-    <StoryPromo image={Image} info={Info} />,
+    <StoryPromo image={Image} info={Info({ topStory: false })} />,
   );
 });
 
 describe('StoryPromo with Media Indicator', () => {
   shouldMatchSnapshot(
     'should render correctly',
-    <StoryPromo image={Image} info={Info} mediaIndicator={mediaInfo} />,
+    <StoryPromo
+      image={Image}
+      info={Info({ topStory: false })}
+      mediaIndicator={mediaInfo}
+    />,
   );
+});
+
+describe('StoryPromo - Top Story', () => {
+  shouldMatchSnapshot(
+    'should render correctly',
+    <StoryPromo image={Image} info={Info({ topStory: true })} topStory />,
+  );
+});
+
+describe('StoryPromo - Top Story with Media Indicator', () => {
+  shouldMatchSnapshot(
+    'should render correctly',
+    <StoryPromo
+      image={Image}
+      info={Info({ topStory: true })}
+      mediaIndicator={mediaInfo}
+      topStory
+    />,
+  );
+});
+
+describe('assertions', () => {
+  it('should render h3, a, p, time', () => {
+    const { container } = render(
+      <StoryPromo
+        image={Image}
+        info={Info({ topStory: true })}
+        mediaIndicator={mediaInfo}
+      />,
+    );
+
+    expect(container.querySelectorAll('h3 a')[0].innerHTML).toEqual(
+      'The headline of the promo',
+    );
+    expect(container.getElementsByTagName('p')[0].innerHTML).toEqual(
+      'The summary of the promo',
+    );
+
+    const time = container.getElementsByTagName('time')[0];
+    const spans = time.getElementsByTagName('span');
+
+    expect(spans[0].innerHTML).toEqual('Video 2 minutes 15 seconds');
+    expect(spans[1].innerHTML).toEqual('2:15');
+
+    const image = container.getElementsByTagName('img')[0];
+
+    expect(image.getAttribute('src')).toEqual('https://foobar.com/image.png');
+    expect(image.getAttribute('alt')).toEqual('Alt text');
+  });
 });