+++ packages/containers/psammead-timestamp-container/src/index.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { number, string, bool, shape } from 'prop-types';
 import { scriptPropType } from '@bbc/gel-foundations/prop-types';
 import Timestamp from '@bbc/psammead-timestamp';
+import moment from 'moment-timezone';
 import {
   isValidDateTime,
   formatUnixTimestamp,
@@ -13,22 +14,29 @@ const TimestampContainer = ({
   dateTimeFormat,
   format,
   isRelative,
+  padding,
   prefix,
   suffix,
   timezone,
   script,
+  locale,
 }) => {
   if (!isValidDateTime(new Date(timestamp))) {
     return null;
   }
 
+  if (locale) {
+    moment.locale(locale);
+  }
+
   return (
     <Timestamp
       datetime={formatUnixTimestamp(timestamp, dateTimeFormat, timezone)}
+      padding={padding}
       script={script}
     >
       {prefix ? `${prefix} ` : null}
-      {showRelativeTime(timestamp, isRelative, format, timezone)}
+      {showRelativeTime(timestamp, isRelative, format, timezone, locale)}
       {suffix ? ` ${suffix}` : null}
     </Timestamp>
   );
@@ -40,17 +48,21 @@ TimestampContainer.propTypes = {
   isRelative: bool,
   format: string,
   timezone: string,
+  padding: bool,
   prefix: string,
   suffix: string,
   script: shape(scriptPropType).isRequired,
+  locale: string,
 };
 
 TimestampContainer.defaultProps = {
   isRelative: false,
   format: null,
   timezone: 'Europe/London',
+  padding: true,
   prefix: null,
   suffix: null,
+  locale: null,
 };
 
 export default TimestampContainer;