+++ packages/components/psammead-media-indicator/CHANGELOG.md
@@ -3,6 +3,9 @@
 <!-- prettier-ignore -->
 | Version | Description |
 | ------- | ----------- |
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0   | [PR#679](https://github.com/BBC-News/psammead/pull/679) Bump version number |
+| 0.2.1   | [PR#644](https://github.com/BBC-News/psammead/pull/644) Fixes for screenreader UX |
 | 0.2.0   | [PR#550](https://github.com/BBC-News/psammead/pull/550) Add audio svg and type prop to switch media icon |
 | 0.1.1   | [PR#515](https://github.com/BBC-News/psammead/pull/515) Update story to use dirDecorator |
 | 0.1.0   | [PR#448](https://github.com/BBC-News/psammead/pull/448) Create initial package. |