+++ packages/components/psammead-section-label/CHANGELOG.md
@@ -3,4 +3,9 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0 | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
+| 0.1.1 | [PR#610](https://github.com/bbc/psammead/pull/610) Add visuallyHidden prop. |
+| 0.1.0 | [PR#616](https://github.com/BBC/psammead/pull/616) Remove alpha tag |
+| 0.1.0-alpha.1 | [PR#611](https://github.com/BBC/psammead/pull/611) Use border top for the horizontal bar |
 | 0.1.0-alpha.0 | [PR#449](https://github.com/BBC/psammead/pull/449) Create initial package with Section Label components. |