+++ packages/components/psammead-sitewide-links/CHANGELOG.md
@@ -3,6 +3,11 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0  | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
+| 0.3.10  | [PR#585](https://github.com/bbc/psammead/pull/604) Remove nanoid package usage for generating list item keys |
+| 0.3.9   | [PR#585](https://github.com/bbc/psammead/pull/585) Make row count dynamic on number of columns and items |
+| 0.3.8   | [PR#576](https://github.com/bbc/psammead/pull/576) Display list in a single column under 240px |
 | 0.3.7   | [PR#498](https://github.com/bbc/psammead/pull/498) Update stories to use new input provider |
 | 0.3.6   | [PR#430](https://github.com/bbc/psammead/pull/430) Add text knobs to story |
 | 0.3.5   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |