+++ packages/components/psammead-timestamp/CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0   | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
+| 0.3.4   | [PR#661](https://github.com/bbc/psammead/pull/661) Update @bbc/gel-foundations to version 3.0.0 |
+| 0.3.3   | [PR#587](https://github.com/bbc/psammead/pull/587) Update colour to C_METAL |
 | 0.3.2   | [PR#515](https://github.com/bbc/psammead/pull/515) Update story to use dirDecorator |
 | 0.3.1   | [PR#493](https://github.com/bbc/psammead/pull/493) Add optional padding prop |
 | 0.3.0   | [PR#494](https://github.com/bbc/psammead/pull/494) Make script a required prop. |